Assistant checkpoint: Remove decimals from loan amount input field
Assistant generated file changes:
- client/src/components/calculators/AdditionalPaymentCalculator.tsx: Remove decimals from loan amount input

---

User prompt:

Remove the decimals in this input field

+++ client/src/components/calculators/AdditionalPaymentCalculator.tsx
@@ -228,9 +228,9 @@ export default function AdditionalPaymentCalculator({ onCalculate }: AdditionalP
                           {...field}
                           className="pl-8"
                           onChange={(e) => {
-                            const value = e.target.value.replace(/[^0-9.]/g, "");
+                            const value = e.target.value.replace(/[^0-9]/g, "");
                             const formattedValue = value
-                              ? formatCurrency(parseFloat(value))
+                              ? formatCurrency(parseInt(value))
                               : "";
                             field.onChange(formattedValue);
                           }}
